Update_2.0.0/Refactor: Divide godlike file
 - Godlike file raid_manager.py divide into manager.py,
registration.py, save_load.py.
 - Realize pattern Singleton for database connection and
Raids list.
 - Move cog.listiner on_reaction_add/on_reaction_remove
into events.py but leaving work with a raid events with
 reactions in manager.py.
